refactor : dto -> vo 이름 변경, application.yml 수정, 예외처리 명확하게 변경
