+++ src/main/java/com/firstticket/bookingservice/booking/domain/service/PaymentOperator.java
@@ -1,6 +1,6 @@
 package com.firstticket.bookingservice.booking.domain.service;
 
-import com.firstticket.bookingservice.booking.domain.service.dto.PaymentResult;
+import com.firstticket.bookingservice.booking.domain.service.vo.PaymentResult;
 import java.util.UUID;
 
 public interface PaymentOperator {
