+++ src/main/java/com/firstticket/bookingservice/booking/domain/service/ProgramOperator.java
@@ -1,6 +1,6 @@
 package com.firstticket.bookingservice.booking.domain.service;
 
-import com.firstticket.bookingservice.booking.domain.service.dto.ProgramScheduleResult;
+import com.firstticket.bookingservice.booking.domain.service.vo.ProgramScheduleResult;
 import java.util.UUID;
 
 public interface ProgramOperator {
