+++ src/main/resources/application.yml
@@ -5,7 +5,7 @@ spring:
   config:
     import:
       - "optional:file:.env[.properties]"
-      - "optional:configserver:${CONFIG_SERVER_URL:http://localhost:8888}"
+      - "optional:configserver:"
 
   cloud:
     config:
@@ -16,5 +16,4 @@ spring:
         service-id: config-server
 
   profiles:
-    active: local
     include: kafka
