+++ src/test/java/com/firstticket/bookingservice/booking/application/BookingCommandServiceTest.java
@@ -14,9 +14,9 @@ import com.firstticket.bookingservice.booking.domain.service.PaymentOperator;
 import com.firstticket.bookingservice.booking.domain.service.ProgramOperator;
 import com.firstticket.bookingservice.booking.domain.service.PublishEvent;
 import com.firstticket.bookingservice.booking.domain.service.SeatOperator;
-import com.firstticket.bookingservice.booking.domain.service.dto.HeldSeatResult;
-import com.firstticket.bookingservice.booking.domain.service.dto.PaymentResult;
-import com.firstticket.bookingservice.booking.domain.service.dto.ProgramScheduleResult;
+import com.firstticket.bookingservice.booking.domain.service.vo.HeldSeatResult;
+import com.firstticket.bookingservice.booking.domain.service.vo.PaymentResult;
+import com.firstticket.bookingservice.booking.domain.service.vo.ProgramScheduleResult;
 import com.firstticket.bookingservice.booking.domain.Booking;
 import com.firstticket.bookingservice.booking.domain.BookingRepository;
 import com.firstticket.bookingservice.booking.domain.BookingStatus;
